// app/types/geoData.ts
export interface CityGeoCoding {
  id: string
  name: string
  latitude: number
  longitude: number
  elevation: number
  feature_code: string
  country_code: string
  admin1_id: string
  admin2_id: string
  admin3_id: string
  admin4_id: string
  timezone: string
  population: number
  postcodes: string[]
  country_id: string
  country: string
  admin1: string
  admin2: string
  admin3: string
  admin4: string
}